requirements.txt
Django~=3.1.3
djangorestframework~=3.12.2
psycopg2-binary==2.8.6
django-elasticsearch~=7.0.2
elasticsearch~=7.10.1